feat: reset admin password

+++ src/main/java/danang03/STBackend/domain/member/MemberRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MemberRepository extends JpaRepository<Member, Integer>, PagingAndSortingRepository<Member, Integer> {
+public interface MemberRepository extends JpaRepository<Member, Long>, PagingAndSortingRepository<Member, Long> {
     Optional<Member> findByEmail(String email);
 
     // 이메일 기준으로 중복 여부를 확인하기 위해
